added mechanism for a Connection's fifo to not be monitored while it is waiting to connect to a peer
This commit lays some foundations for the upcoming implementation of encryption
in Cryptocomms. When encryption is added, two Cryptocomms hosts will need to
establish a connection before any data can be sent. While this connection is being
made, there is no point in monitoring the Connection object's fifo for user data,
since there is nothing that can be done with it. This commit implements this, but
due to the lack of the rest of the encryption logic it does not really do anything
yet.

Theses changes cause one test in Connection.tests.cpp to fail due to an intentional
change in the behaviour of Connnection. This is best resolved by implementing the full
encryption logic, which will be done in the next few commits, so the test has been
temporarily disabled.