created new mechanism for more secure segment number storage
Before this commit, SegmentNumGenerator stored its record of which
segment numbers had been used in a single file. If this file became
corrupt (eg because Cryptocomms was killed while writing it), then
Cryptocomms had no way to recover by itself. This commit changes
SegmentNumGenerator to use copies of the file for storage,
significantly reducing the likelihood of unrecoverable corruption
of the record.